Extract slot_available/2 from Waiver.Validate
* Extract slot_available/2 from Waiver.Validate
* Plan to use slot_available to validate draft picks
* Moved to ValidateHelper module
* See #461
