+++ lib/ex338/waiver/validate.ex
@@ -3,7 +3,7 @@ defmodule Ex338.Waiver.Validate do
 
   use Ex338Web, :model
 
-  alias Ex338.{FantasyPlayer, FantasyTeam, RosterPosition, Repo}
+  alias Ex338.{FantasyPlayer, FantasyTeam, RosterPosition, Repo, ValidateHelpers}
 
   def add_or_drop(waiver_changeset) do
     add_player = fetch_change(waiver_changeset, :add_fantasy_player_id)
@@ -87,15 +87,6 @@ defmodule Ex338.Waiver.Validate do
         |> RosterPosition.count_positions_for_team(team_id)
         |> do_open_position(waiver_changeset)
     end
-  end
-
-  def slot_available?(roster_positions, max_flex_spots) do
-    total_slot_count = count_total_slots(roster_positions)
-
-    roster_positions
-    |> count_regular_slots
-    |> calculate_flex_slots_used(total_slot_count)
-    |> compare_flex_slots(max_flex_spots)
   end
 
   def wait_period_open(waiver_changeset) do
@@ -198,7 +189,7 @@ defmodule Ex338.Waiver.Validate do
   end
 
   defp do_max_flex_slots(waiver_changeset, future_positions, max_flex_spots) do
-    case slot_available?(future_positions, max_flex_spots) do
+    case ValidateHelpers.slot_available?(future_positions, max_flex_spots) do
       true ->
         waiver_changeset
 
@@ -209,27 +200,6 @@ defmodule Ex338.Waiver.Validate do
           "No flex position available for this player"
         )
     end
-  end
-
-  ## slot_available?
-
-  defp count_total_slots(slots) do
-    Enum.count(slots)
-  end
-
-  defp count_regular_slots(slots) do
-    slots
-    |> Enum.map(& &1.fantasy_player.sports_league_id)
-    |> Enum.uniq()
-    |> Enum.count()
-  end
-
-  defp calculate_flex_slots_used(regular_slots_filled, total_filled) do
-    total_filled - regular_slots_filled
-  end
-
-  defp compare_flex_slots(num_filled, max_flex_spots) do
-    num_filled <= max_flex_spots
   end
 
   ## wait_period_open
